refactor: student view mobile audit, component reorganization, and profile inline
- Reorganize student components into info/ and tracker/ subdirectories
- Inline StudentInfo into profile/page.tsx and rebuild as a full editable form
- Rename StatusComponent to RegistrationStatus
- Add formatAddress utility and replace ad-hoc address formatting across the codebase
- Change student dashboard layout breakpoints from 2xl to md for better mobile support
- Modernize Tailwind: flex-shrink-0 → shrink-0, w-[var(--...)] → w-(--...)

Closes #421
