+++ frontend/src/app/(student)/_components/PartyRegistrationForm.tsx
@@ -147,7 +147,7 @@ export default function PartyRegistrationForm({
     placeId: string;
   } | null>(null);
 
-  const handleSubmit = async (e: React.FormEvent) => {
+  const handleSubmit = async (e: React.SyntheticEvent<HTMLFormElement>) => {
     e.preventDefault();
     setErrors({});
