+++ frontend/src/app/(student)/_components/tracker/DeletePartyDialog.tsx
