+++ frontend/src/app/(student)/_components/tracker/RegistrationIncidentCard.tsx
