+++ frontend/src/app/(student)/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
-import RegistrationTracker from "@/app/(student)/_components/RegistrationTracker";
-import StatusComponent from "@/app/(student)/_components/StatusComponent";
+import RegistrationStatus from "@/app/(student)/_components/RegistrationStatus";
+import RegistrationTracker from "@/app/(student)/_components/tracker/RegistrationTracker";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useCurrentStudent } from "@/lib/api/student/student.queries";
-import { isFromThisSchoolYear } from "@/lib/utils";
+import { formatAddress, isFromThisSchoolYear } from "@/lib/utils";
 import { AlertTriangleIcon, Info } from "lucide-react";
 import Link from "next/link";
-import PartyRegistrationInfo from "./_components/PartyRegistrationInfo";
-import PartySmartInfo from "./_components/PartySmartInfo";
+import PartyRegistrationInfo from "./_components/info/PartyRegistrationInfo";
+import PartySmartInfo from "./_components/info/PartySmartInfo";
 
 export default function StudentDashboard() {
   const studentQuery = useCurrentStudent();
@@ -19,8 +19,8 @@ export default function StudentDashboard() {
   const location = studentQuery?.data?.residence?.location;
 
   return (
-    <div className="px-4 sm:px-8 pt-6 pb-6 flex flex-col 2xl:flex-row gap-4 max-w-4xl mx-auto w-full 2xl:max-w-11/12 2xl:gap-24 h-full overflow-hidden min-h-0">
-      <div className="2xl:w-1/2 flex flex-col flex-1 2xl:flex-none min-h-0">
+    <div className="px-4 sm:px-8 pt-6 pb-6 flex flex-col md:flex-row gap-4 max-w-4xl mx-auto w-full md:max-w-11/12 md:gap-24 h-full overflow-hidden min-h-0">
+      <div className="md:w-1/2 flex flex-col flex-1 md:flex-none min-h-0">
         <div className="flex items-center justify-between">
           <h1 className="page-title w-1/2">Parties</h1>
           <div className="content text-wrap text-end w-1/2">
@@ -30,8 +30,11 @@ export default function StudentDashboard() {
               ) : (
                 validResidence && (
                   <p>
-                    {location?.street_number} {location?.street_name}{" "}
-                    {location?.unit}
+                    {formatAddress(location, [
+                      "street_number",
+                      "street_name",
+                      "unit",
+                    ])}
                   </p>
                 )
               )}
@@ -44,12 +47,12 @@ export default function StudentDashboard() {
           <h2 className="page-title mb-2">Registration Status </h2>
           <Link
             href="/about-party-registration"
-            className="content flex items-center mb-2 2xl:hidden"
+            className="content flex items-center mb-2 md:hidden"
           >
-            <Info className="h-4 w-4 inline-block mr-1" />
+            <Info className="size-4 inline-block mr-1" />
             <p className="underline">Learn About Party Registration</p>
           </Link>
-          <StatusComponent
+          <RegistrationStatus
             last_registered={studentQuery.data?.last_registered}
             hold_expiration={
               studentQuery.data?.residence?.location.hold_expiration
@@ -59,9 +62,10 @@ export default function StudentDashboard() {
         </div>
       </div>
 
-      <div className="hidden 2xl:flex 2xl:flex-col 2xl:w-1/2 2xl:justify-between">
+      <div className="hidden md:flex md:flex-col md:w-1/2 ">
         <PartyRegistrationInfo />
-        <div className="flex flex-col items-center text-center content">
+        <div className="flex-1" />
+        <div className="flex flex-col items-center text-center content [@media(max-height:725px)]:hidden">
           <AlertTriangleIcon />
           <p className="max-w-md">
             Keep in mind that the party registration program only pertains to
@@ -69,6 +73,7 @@ export default function StudentDashboard() {
             likely result in local law enforcement showing up without a warning.
           </p>
         </div>
+        <div className="flex-1" />
         <PartySmartInfo />
       </div>
     </div>
