+++ frontend/src/app/staff/_components/account/PoliceAccountForm.tsx
@@ -51,7 +51,7 @@ export default function PoliceAccountForm({
   const [errors, setErrors] = useState<Record<string, string>>({});
   const [isSubmitting, setIsSubmitting] = useState(false);
 
-  const handleSubmit = async (e: React.FormEvent) => {
+  const handleSubmit = async (e: React.SyntheticEvent<HTMLFormElement>) => {
     e.preventDefault();
     setErrors({});
