+++ frontend/src/app/staff/_components/incident/IncidentTableForm.tsx
@@ -121,7 +121,7 @@ export default function IncidentTableForm({
     }
   };
 
-  const handleSubmit = async (e: React.FormEvent) => {
+  const handleSubmit = async (e: React.SyntheticEvent<HTMLFormElement>) => {
     e.preventDefault();
     setErrors({});
